🐛 fix: fix display error when using DEFAULT_AGENT_CONFIG env
(cherry picked from commit e46215c17a8761d27c1194416a4c476b9e13d2bd)
