+++ src/server/mock.ts
@@ -1,0 +1,8 @@
+/**
+ * This file contains the root router of your tRPC-backend
+ */
+import { createCallerFactory } from '@/libs/trpc';
+
+import { appRouter } from './routers';
+
+export const createCaller = createCallerFactory(appRouter);
