+++ src/store/agent/initialState.ts
@@ -1,7 +1,7 @@
-import { AgentState, initialSessionState } from './slices/chat/initialState';
+import { AgentState, initialAgentChatState } from './slices/chat/initialState';
 
 export type SessionStoreState = AgentState;
 
 export const initialState: SessionStoreState = {
-  ...initialSessionState,
+  ...initialAgentChatState,
 };
