+++ src/store/agent/slices/chat/initialState.ts
@@ -1,14 +1,20 @@
+import { DeepPartial } from 'utility-types';
+
 import { DEFAULT_AGENT_CONFIG } from '@/const/settings';
 import { LobeAgentConfig } from '@/types/agent';
 
 export interface AgentState {
   activeId: string;
-  agentConfig: LobeAgentConfig;
+  agentConfig: DeepPartial<LobeAgentConfig>;
+  defaultAgentConfig: LobeAgentConfig;
   isAgentConfigInit: boolean;
+  isDefaultAgentConfigInit: boolean;
 }
 
-export const initialSessionState: AgentState = {
+export const initialAgentChatState: AgentState = {
   activeId: 'inbox',
-  agentConfig: DEFAULT_AGENT_CONFIG,
+  agentConfig: {},
+  defaultAgentConfig: DEFAULT_AGENT_CONFIG,
   isAgentConfigInit: false,
+  isDefaultAgentConfigInit: false,
 };
